Add authService with login and logout
Move login from component to the new service
